O7E-50 Refactor VehicleSummary import path to use vehicleDataModels; remove obsolete VehicleSummary interface from models.ts
